Function rslogo::run

pub fn run(
    file_content: String,
    image: &mut Image,
    width: u32,
    height: u32
) -> Result<(), String>

RSLogo

Parsing RSLogo Scripts and Drawing in Images

Arguments

file_content - A string containing the RSLogo script to be executed.

image - A mutable reference to an usvg::Image object where the drawing will be rendered.

height - The height of the output image in pixels.

width - The width of the output image in pixels.

Returns

returns Ok(())

If the RSLogo scripts executes successfully, this function will drawing the specified graphics on the image.

returns Err(String)

If there is an error during parsing, analyzing, or executing the script, this function will return Err(String) with the String providing details about the nature of the error

Examples

use unsvg::Image;
use rslogo::run;
let file_content = String::from("PENUP\nFORWARD \"50\nPENDOWN\nFORWARD \"50"); // Example script
let mut image = Image::new(200, 200); // Placeholder for the actual image object
let result = run(file_content, &mut image, 200, 200);
assert!(result.is_ok());